requirements.txt
Django==2.0.7
psycopg2-binary==2.7.5
pytz==2018.5
django-registration==2.4.1
